Spawn Light Feature
Attaching the Light Spawner script to the Light Spawner object in scene
10 and giving it the sphere prefab well set up the functionality to
spawn lights. When you want to, call the Spawn light function from the
Light spawner class and it will handle the rest. The object that is
being spawned must have a circle behaviour script attached to it, if it
does not, then the code will add one for you.